Integer promotion fix in simtypes.h
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@3161 8aca7d54-2c30-db11-9de9-000461428c89

+++ simtypes.h
@@ -143,7 +143,8 @@ inline uint16 endian_uint16(const uint16 *d)
 	return  *d;
 #else
 	const uint8 *data = (const uint8 *)d;
-	return ((uint16)data[0]) | ( ((uint16)data[1]) << 8 );
+	/* according to C standard uint8 will be anyway extended to unsigned */
+	return (uint16)(data[0] | ( data[1] << 8 ));
 #endif
 }
